Move fixtures to conftest
Creating a single file with all fixtures.
Moving all the fixtures to this file.

close #43
