+++ users/tests.py
@@ -2,54 +2,6 @@ import pytest
 from users.models import User
 
 
-@pytest.fixture
-def user_1() -> User:
-    user1 = User.objects.create_user(username='user1',
-                                     first_name='first_name',
-                                     last_name='last_name',
-                                     email='email',
-                                     password='password1')
-    user1.role = 'Student'
-    user1.save()
-    return user1
-
-
-@pytest.fixture
-def user_2() -> User:
-    user2 = User.objects.create_user(username='user2',
-                                     first_name='first_name',
-                                     last_name='last_name',
-                                     email='email',
-                                     password='password2')
-    user2.role = 'Student'
-    user2.save()
-    return user2
-
-
-@pytest.fixture
-def create_user_student() -> User:
-    user_student = User.objects.create_user(username='user_student',
-                                            first_name='first_name',
-                                            last_name='last_name',
-                                            email='email',
-                                            password='password1')
-    user_student.role = 'Student'
-    user_student.save()
-    return user_student
-
-
-@pytest.fixture
-def create_user_HR() -> User:
-    user_HR = User.objects.create_user(username='user_HR',
-                                       first_name='first_name',
-                                       last_name='last_name',
-                                       email='email',
-                                       password='password1')
-    user_HR.role = 'HR'
-    user_HR.save()
-    return user_HR
-
-
 @pytest.mark.django_db
 def test_create_user(user_1):
     assert isinstance(user_1, User)
